Embeded checkpoint
