+++ Book Logger/Base.lproj/Main.storyboard
@@ -66,10 +66,10 @@
             </objects>
             <point key="canvasLocation" x="420" y="-666.56671664167925"/>
         </scene>
-        <!--Photo View-->
+        <!--Page Controller-->
         <scene sceneID="Vfk-Bb-fu1">
             <objects>
-                <viewController id="nAh-gS-HXs" customClass="PhotoView" customModule="Book_Logger" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="nAh-gS-HXs" customClass="PageController" customModule="Book_Logger" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="VnS-Du-Mj0">
                         <rect key="frame" x="0.0" y="0.0" width="311" height="478"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -83,10 +83,9 @@
                             <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="FBF-Lv-ql5">
                                 <rect key="frame" x="16" y="44" width="279" height="169"/>
                             </imageView>
-                            <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" textAlignment="natural" translatesAutoresizingMaskIntoConstraints="NO" id="iQT-bR-9JR">
+                            <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" text="Scanning image now..." textAlignment="natural" translatesAutoresizingMaskIntoConstraints="NO" id="iQT-bR-9JR">
                                 <rect key="frame" x="16" y="264.5" width="279" height="193.5"/>
                                 <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                <string key="text">Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.</string>
                                 <fontDescription key="fontDescription" type="system" pointSize="14"/>
                                 <textInputTraits key="textInputTraits" autocapitalizationType="sentences"/>
                             </textView>
@@ -134,10 +133,10 @@
             </objects>
             <point key="canvasLocation" x="1170" y="-721"/>
         </scene>
-        <!--Collection View Controller-->
+        <!--Books Controller-->
         <scene sceneID="lve-Kv-K2N">
             <objects>
-                <collectionViewController id="L3g-6B-KKO" customClass="CollectionViewController" customModule="Book_Logger" customModuleProvider="target" sceneMemberID="viewController">
+                <collectionViewController id="L3g-6B-KKO" customClass="BooksController" customModule="Book_Logger" customModuleProvider="target" sceneMemberID="viewController">
                     <collectionView key="view" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="prototypes" id="FCo-eg-WyK">
                         <rect key="frame" x="0.0" y="0.0" width="311" height="478"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -230,10 +229,10 @@
             </objects>
             <point key="canvasLocation" x="-375.19999999999999" y="-666.56671664167925"/>
         </scene>
-        <!--Collection View Controller-->
+        <!--Pages Controller-->
         <scene sceneID="CRF-Gp-npF">
             <objects>
-                <collectionViewController id="JjE-og-WjJ" customClass="CollectionViewController" customModule="Book_Logger" customModuleProvider="target" sceneMemberID="viewController">
+                <collectionViewController id="JjE-og-WjJ" customClass="PagesController" customModule="Book_Logger" customModuleProvider="target" sceneMemberID="viewController">
                     <collectionView key="view" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="prototypes" id="Rq2-g5-Kpg">
                         <rect key="frame" x="0.0" y="0.0" width="311" height="478"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -257,7 +256,7 @@
                                             <fontDescription key="fontDescription" type="system" pointSize="50"/>
                                             <state key="normal" title="+"/>
                                             <connections>
-                                                <segue destination="nAh-gS-HXs" kind="show" id="FdE-Zy-4AK"/>
+                                                <action selector="newPage:" destination="JjE-og-WjJ" eventType="touchUpInside" id="KwN-I8-JhP"/>
                                             </connections>
                                         </button>
                                     </subviews>
@@ -275,6 +274,9 @@
                             <outlet property="delegate" destination="JjE-og-WjJ" id="Eim-Ku-XCm"/>
                         </connections>
                     </collectionView>
+                    <connections>
+                        <segue destination="nAh-gS-HXs" kind="show" identifier="toPage" id="nCO-Ry-MPx"/>
+                    </connections>
                 </collectionViewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="1sl-W6-Xum" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
