+++ Book Logger/CollectionViewController.swift
@@ -1,24 +1,19 @@
-//
-//  CollectionViewController.swift
-//  Book Logger
-//
-//  Created by Mac on 9/30/18.
-//  Copyright © 2018 NC. All rights reserved.
-//
+
 
 import UIKit
 
 private let reuseIdentifier = "Cell"
 
 class CollectionViewController: UICollectionViewController, UICollectionViewDelegateFlowLayout {
+    
+    var images:[UIImage]!
+    var titles:[String]!
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
         // Register cell classes
         self.collectionView!.register(UICollectionViewCell.self, forCellWithReuseIdentifier: reuseIdentifier)
-
-        // Do any additional setup after loading the view.
     }
 
     override func numberOfSections(in collectionView: UICollectionView) -> Int {
@@ -27,22 +22,7 @@ class CollectionViewController: UICollectionViewController, UICollectionViewDele
 
 
     override func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        // Return the number of current books/pages, plus 1 for the 'new image' button
         return 1
-    }
-
-    
-    override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "bookCell", for: indexPath)
-    
-        // Configure the cell
-        cell.layer.borderWidth = 1
-        cell.layer.borderColor = UIColor.lightGray.cgColor
-        
-        // Use this to load the image page
-        print(indexPath)
-    
-        return cell
     }
 
     func collectionView(_ collectionView: UICollectionView,
