+++ Book Logger/PageController.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+class PageController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+}
